# DOT-LMS-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DOT-LMS API Documentation",
        "contact": {
            "email": "ruthgetaneh5@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "DOT-LMS API Server"
        }
    ],
    "paths": {
...
        "/Student": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "Get list of Students in the DOT-LMS",
                "description": "Returns list of Students",
                "operationId": "38e52c1874d60354b8bf504acb766db3",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Student"
                ],
                "summary": "Store new Student's data into the database",
                "description": "Returns the entered Student's data",
                "operationId": "07aa417aa0b004fe04713b8bf8d84dc9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StudentUser"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/Student/AddbyCSV": {
            "post": {
                "tags": [
                    "Student"
                ],
                "summary": "Store",
                "description": "Returns an array of the file name and files",
                "operationId": "281b080c11d64e3d874995a5558fab1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "csv_file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/Student/{id}": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "Get a specific Student",
                "description": "Returns the specified Student's data",
                "operationId": "6048f0228dbb1745eefeab92e5cc6fb5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Student's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "DBUR-6832-23"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Student"
                ],
                "summary": "Update existing Student",
                "description": "Returns updated Student's data",
                "operationId": "1096a2fd3de535d2e25528a9276494c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Student's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "DBUR-6832-23"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StudentUser"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Student"
                ],
                "summary": "Delete existing Student",
                "description": "Deletes a record and returns no content",
                "operationId": "075af06c7b9368bae4eea1b537cf47ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Student's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "DBUR-6832-23"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/Student/Assign_courses": {
            "post": {
                "tags": [
                    "Student"
                ],
                "summary": "Assign courses to students",
                "description": "Assign courses to students",
                "operationId": "b694b1fe3317f4080d08cb9f481c369e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "student_id": {
                                        "type": "string",
                                        "example": "DBU-3213-23"
                                    },
                                    "course_id": {
                                        "type": "string",
                                        "example": "CUS-3245"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/Student/AssignedCourse/{student_id}": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "Get assigned courses for a student",
                "description": "Returns the assigned courses for a student",
                "operationId": "65ed7cec561893838e0263f79d2797ea",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "path",
                        "description": "Student's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "DBU-6832-23"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/Student/AssignbyBatch": {
            "post": {
                "tags": [
                    "Student"
                ],
                "summary": "Assign courses to students by Batch",
                "description": "Assign courses to students by Batch",
                "operationId": "4c8fb0953af65fd80e47316c07c4557a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "entry_year": {
                                        "type": "string",
                                        "example": "23"
                                    },
                                    "course_id": {
                                        "type": "string",
                                        "example": "CUS-3245"
                                    },
                                    "department": {
                                        "type": "string",
                                        "example": "computer science"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/Student/Course/{course_id}": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "Get a list of students in a specified course",
                "operationId": "7143ff293b1ff2b48dac5ac0b47902a9",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "description": "Course's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "CUS-6832"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StudentUser": {
                "title": "Student",
                "description": "Student data",
                "properties": {
                    "first_name": {
                        "title": "first_name",
                        "description": "First Name of the Student",
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "title": "last_name",
                        "description": "last Name of the Student",
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "title": "email",
                        "description": "email address of the Student",
                        "type": "string",
                        "example": "Johndoe@gmail.com"
                    },
                    "department": {
                        "title": "department",
                        "description": "department of the Student",
                        "type": "string",
                        "example": "Software Engineering"
                    },
                    "year": {
                        "title": "year",
                        "description": "Academic year of the Student",
                        "type": "string",
                        "example": "4"
                    },
                    "semester": {
                        "title": "semester",
                        "description": "in which semester of the academic year is the student currently attending",
                        "type": "string",
                        "example": "2"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "apiKey",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "api_key",
                "in": "header"
            }
        }
    },
...
}